// administration_management-client_side/src/interfaces/Activity.ts
export default class Activity {
    activityId: number = 0;
    communityAdministrationId: string = '';
    activityName: string = "";
    description: string = "";
    cost: number = 0;
    targetAudience: string = "";
    endTimeOfPreview: Date = new Date();
    maxParticipants: number = 0;
    day: string = "";
    beginngTime: string = "";
    endTime: string = "";
    date: string = "";
}